+++ app/data/spells/fire.ts
@@ -1,4 +1,5 @@
 import { Spell } from "../../types/spell.ts";
+import { APP_DOMAIN } from '../../config/index.ts';
 
 export const fireSpell: Array<Spell> = [
   {
@@ -7,6 +8,7 @@ export const fireSpell: Array<Spell> = [
       level: 1,
       effect: "Increases the attack skill of target (or all) " +
         "allied creature for melee attacks",
+      imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Bloodlust.png`,
   },
   {
     id: "2",
@@ -14,6 +16,7 @@ export const fireSpell: Array<Spell> = [
     level: 1,
     effect: "The targeted (or all) enemy unit(s) deliver" +
       "minimum or reduced damage when attacking",
+    imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Curse.png`,
   },
   {
     id: "3",
@@ -23,6 +26,7 @@ export const fireSpell: Array<Spell> = [
       "Basic: Single ally. Expert: All allies. Effect " +
       "increased from 30%-30%-50%-50% to 50%-50%-75%-75%. " +
       "Horn of the Abyss",
+    imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Protection_from_Fire.png`,
   },
   {
     id: "4",
@@ -30,6 +34,7 @@ export const fireSpell: Array<Spell> = [
     level: 2,
     effect: "Prevents any action from being taken by " +
       "target enemy creature. ",
+    imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Blind.png`,
   },
   {
     id: "5",
@@ -38,12 +43,14 @@ export const fireSpell: Array<Spell> = [
     effect: "Creates wall of fire at target location " +
       "that inflicts magic damage on creatures that" + 
       "pass through",
+    imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Fire_Wall.png`,
   },
   {
     id: "6",
     name: "Fireball",
     level: 3,
     effect: "Inflicts magic damage on all creatures in area",
+    imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Fireball.png`,
   },
   {
     id: "7",
@@ -51,12 +58,14 @@ export const fireSpell: Array<Spell> = [
     level: 3,
     effect: "locations on the battlefield. Only visible " +
       "to caster and creatures native to the terrain",
+    imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Land_Mine.png`,
   },
   {
     id: "8",
     name: "Misfortune",
     level: 3,
     effect: "Reduces luck. Basic: Single enemy. Expert: All enemies",
+    imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Misfortune.png`,
   },
   {
     id: "9",
@@ -65,12 +74,14 @@ export const fireSpell: Array<Spell> = [
     effect: "Inflicts magic damage on all creatures on the battlefield. " +
       "Damage reduced from 50*(spell power) + 30/30/60/120 to 40*(spell power) " + 
       "+ 30/30/60/120. Horn of the Abyss",
+    imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Armageddon.png`,
   },
   {
     id: "10",
     name: "Berserk",
     level: 4,
     effect: "Forces enemy to attack closest creature",
+    imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Berserk.png`,
   },
   {
     id: "11",
@@ -78,18 +89,21 @@ export const fireSpell: Array<Spell> = [
     level: 4,
     effect: "Inflicts magic damage proportional to " +
       "damage taken from hand-to-hand attacks",
+    imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Fire_Shield.png`,
   },
   {
     id: "12",
     name: "Frenzy",
     level: 4,
     effect: "Increases attack rating and sets defense rating to 0",
+    imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Frenzy.png`,
   },
   {
     id: "13",
     name: "Inferno",
     level: 4,
     effect: "Inflicts magic damage on all creatures in area",
+    imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Inferno.png`,
   },
   {
     id: "14",
@@ -98,6 +112,7 @@ export const fireSpell: Array<Spell> = [
     effect: " 	Increases attack rating of a single ally against " + 
       "Dragons, Behemoths and Hydras. Advanced: Effect applies" + 
       "to Angels and Devils. Expert: Effect applies to Titans",
+    imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Slayer.png`,
   },
   {
     id: "15",
@@ -105,11 +120,13 @@ export const fireSpell: Array<Spell> = [
     level: 5,
     effect: "Destroys and removes living, unkilled allies to bring" +
       "previously living, killed allies back to life",
+    imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Sacrifice.png`,
   },
   {
     id: "16",
     name: "Summon Fire Elemental",
     level: 5,
     effect: "Summons allied Fire Elementals for the duration of combat",
+    imageUrl: `${APP_DOMAIN}assets/images/spell/fire/Summon_Fire_Elemental.png`,
   },
 ];